requirements.txt
# Azure Functions
azure-functions

# HTTP requests
requests
